requirements.txt
langchain==0.3.26
flask==3.1.1 
sentence-transformers==4.1.0 # open source model for embeddings.presently using 'all-MiniLM-L6-v2'
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-pinecone==0.2.8 # Pinecone vector store integration
langchain-community==0.3.26
ipykernel==6.25.2 # Jupyter kernel for running notebooks
langchain-google-genai
markdown==3.4.4 # For rendering markdown in responses
BeautifulSoup4==4.12.2 # For parsing HTML content
#-e .